Greatly simplify the nameserver settings
This commit makes the method of changing a resolvers nameservers more
functional and less buggy.  For example, using a hostname rather than
ip for setting the nameserver is no longer buggy (the previous
implentation would result in preserving the default nameserver and
would append the inspected form of the IPAddr object of the new nameserver)

This commit consolidates the location of changing @config[:nameservers]
into a single location.